Move the check for a valid form element inside init
The conditional check for a non-null form element was relocated from the createFormEventListener method to inside the init method. Since the check was being made in the former, the createWindowEventListener method was still being fired, resulting in the onbeforeleave listener to be executed, even on pages that did not have a form.